Gruntwork + Back-end start
I have filled out some more components, opting for a design first and build the functionality later approach. I got the chats and chat messages and related containers styled nicely and once i finished with most of these components i moved onto getting firebase set up.

This is where i encountered my first major bug that im still struggling with now, which is that i am getting an error related to the cloud storage. It is my first time playing around with cloud storage and i am getting an error to say that the image does not exist in storage, implying i am trying to access it before it is available, which i am not. At this time i have no clue what is causing this i believe it may be something to do with the asynchronous nature of firebase cloud storage. Everything does work as intended on the firebase side the authentication works the storing of the images in cloud works and the saving users to a separate database also works and yet the error persists.