Bug fixed
So after some testing today i realised the issue i was having with regards to the firebase error that the "object could not be found in storage" was happening because the function firebase provides called "getDownloadURL(uploadTask.snapshot.ref)" was not waiting for the saving of the image in storage before trying to read it. I dont believe this is expected to happen as my code is more or less the same as the example given on firebase. Regardless to fix it i added a simple check to make sure the snapshot.metadata exists as it will only exist after the save is complete thus delaying the reading of the file until its done saving in storage.