Finished Authentication
I have finished up all the features that go along with authentication. Making the user able to sign up, login, sign out and navigate between them thanks to react-router's useNavigate hook.

Firebase makes logging in and out of an account really easily simply requiring 2 functions which are signOut and signInWithEmailAndPassword which we pass our auth into and then simply navigate the user accordingly.

I created an AuthContext as it was clear the currently logged in user would need to be accessed all over the application. Given this is simply a single object redux will not be necessary in this project.

I have also added the ability to search for a user by their display name and show them with their image and name, retrieved from the database using firebase query.