Functionality
I spent all day yesterday trying to finish out most of the functionality for the app.

Firstly, i finished implementing the search feature, allowing the logged in user to search for other users to message with. When a user is authenticated i am adding all details and photos excluding the password to a firestore database so that they can be searched using firebase query later. Then if there is a user chat already we fetch them and if not we create one for the 2 users storing it under their combined uid's for authenticating later that the chat being pulled is definitely the chat between the right 2 users. other users can never be shown the wrong chat e.g. between 2 different users because their calculated combined id will never match the one for any other chat than the right one.

Next i created a chat context for storing the second user of each chat and getting the chatId of each chat anywhere in the application to avoid prop drilling and to allow for the changing of the user.

Then i moved onto the Input component which is the input for the chat message. There were a few considerations that made this rather time consuming, firstly that we needed to display both users name, photo and time of message as well as to be able to send images which would also need to be stored through firebase to remain persistent. I first made the input a controlled component. Then we check if we are sending just text or an image attachment as well. If its just text we simply update the chats collection in firebase with the message which will get an id from uuid library the text, the id of the sender and the timestamp. if the message contains an image we first save that image in firebase cloud storage and then we use the link given back to us from firebase storage to then read and display the image in the chat. Then we update the userChats for both the logged in user and the user they are trying to chat to with the timestamp and last message so we can keep track of which chat was most recently used and sort the sidebar chats component based on that.

The chats component makes use of firebase's onSnapshot method to get a live snapshot of the currently logged in user's chats collection. Here we also use the chatContext to change the user when clicking on a chat.

Then we do the same thing for the messages, getting a live snapshot from firebase and simply rendering that to the screen using a ChatMessage component which takes in the message and and using our 2 contexts figures out if its the message owner that it belongs to or if its been sent to the logged in user from someone else, as we need to style the message accordingly. We then use the useRef hook to create a reference for the latest message which allows us to scroll it into view when a new message is sent.

This is more or less all the base features and functionality of a chat application. You can sign up, login search users, send messages and images to users and logout i tried to use a lot of the react tools and features that i have been learning over these last few months into this project to help me in learning them.

Still left to finish:
-Responsiveness
-Friends List or adding users
-Sending attachments other than images
-(maybe) video chat
-(maybe) emoji options

Once these are implemented and the application design is polished off it will be ready for deployment.