Friends list
Added: For the rest of the day today i have coded up the ability to add friends. and have a friends list that is visible for the user and allows them to open up a chat with any friend they like.

There were a few considerations in my approach. Firstly i wanted a scaleable solution and so i have a new firebase collection for friends that are stored under the uid of each user, this does lead to some data redundancy and multiple saves of the same data however it provides us with real-time updates so if a user changes their profile picture, it will also change on the friends list for anyone who has added them as a friend.

Right now both users must add each other to appear on their friends list, however im not yet sure whether i want to have one user need to accept a friend request or whether to simply add the friend for both users when one does it. Given this is just a personal project i will likely go with the latter although if i were making this for a real world app i would definitely make it be requested and accepted before becoming friends.