Responsiveness
Today i have finished out what is more or less the final piece of the puzzle, which is making this responsive. Based on the most popular mobile resolutions i have opted to make this responsive down to 320px max-width. I have implemented a mobile UI which is a little more limited as it does not utilize the friends list.

I have also implemented a burger menu component for the mobile layout and added a small change to the chats component to make it so that if the text is longer than 70 characters it will only take 70 characters and add a "..." to the end to indicate the message continues, rather than displaying a very long message.