Bug Fixes / UI Enhancements
Today after deploying the application i came across a few bugs and things that should be fixed.

Firstly the friends list was not working at all. This was due to me having no code to create a Doc in firebase if one was not found and so it was trying to update a doc that does not exist. It is now conditional so if the Doc is there it will update and if not it will create a new one.

There was another issue with the friends list where when adding a new friend it was adding all of them again thus rendering the same original friends twice causing multiple issues. I fixed this with a conditional that i feel is a bit performance heavy that could be done better but at this time i cant think of the optimal solution so i run a find method to check if the friend we are trying to add is already in there if not we add it, otherwise we return the original friends list.

I added some validation for the input as i noticed that i was able to send empty messages which would still result in a sent message that would be stored even though it has no content. The function will now return if there is no text via a guard clause.

Finally today i added the functionality for the correct date formatting, so if it is today the time of the message is shown, if it is longer than 24 hours it will show the date instead.